+++ src/crontab/src/Mutex/RedisServerMutex.php
@@ -14,6 +14,9 @@ namespace Hyperf\Crontab\Mutex;
 use Hyperf\Crontab\Crontab;
 use Hyperf\Redis\RedisFactory;
 use Hyperf\Utils\Arr;
+use Hyperf\Utils\Coordinator\Constants;
+use Hyperf\Utils\Coordinator\CoordinatorManager;
+use Hyperf\Utils\Coroutine;
 
 class RedisServerMutex implements ServerMutex
 {
@@ -49,6 +52,10 @@ class RedisServerMutex implements ServerMutex
         $result = (bool) $redis->set($mutexName, $this->macAddress, ['NX', 'EX' => $crontab->getMutexExpires()]);
 
         if ($result === true) {
+            Coroutine::create(function() use ($crontab, $redis, $mutexName) {
+               $exited = CoordinatorManager::until(Constants::WORKER_EXIT)->yield($crontab->getMutexExpires());
+               $exited && $redis->del($mutexName);
+            });
             return $result;
         }