+++ src/load-balancer/src/AbstractLoadBalancer.php
@@ -66,8 +66,8 @@ abstract class AbstractLoadBalancer implements LoadBalancerInterface
             $nodes = call($callback);
             is_array($nodes) && $this->setNodes($nodes);
         });
-        Coroutine::create(function () use ($timerId) {
-            CoordinatorManager::get(Constants::ON_WORKER_EXIT)->yield();
+        Coroutine::create(function() use ($timerId){
+            CoordinatorManager::until(Constants::WORKER_EXIT)->yield();
             Timer::clear($timerId);
         });
     }